Rename classes
+ Rename some classes
+ Update "infos.txt" file (provide an example of usage)
