+++ G07/Core/Link.java
@@ -1,0 +1,13 @@
+package Core;
+
+/**
+ * @author JackassDestroyer
+ * Simple link between two nodes (shortcut class)
+ */
+public class Link extends AbstractLink<Node>
+{
+	public Link(Node InSourceNode, Node InDestinationNode) 
+	{
+		super(InSourceNode, InDestinationNode);
+	}
+}
