+++ G07/Core/Node.java
@@ -1,0 +1,8 @@
+package Core;
+
+/**
+ * 
+ * @author JackassDestroyer
+ * Simple graph node (shortcut class)
+ */
+public class Node extends AbstractNode<Link> {}
