--- G07/Core/SimpleLink.java
@@ -1,13 +1,0 @@
-package Core;
-
-/**
- * @author JackassDestroyer
- * @brief Simple link between two nodes, 0 informations added
- */
-public class SimpleLink extends AbstractLink<SimpleNode>
-{
-	public SimpleLink(SimpleNode InSourceNode, SimpleNode InDestinationNode) 
-	{
-		super(InSourceNode, InDestinationNode);
-	}
-}
