--- G07/Core/SimpleNode.java
@@ -1,6 +1,0 @@
-package Core;
-
-public class SimpleNode extends AbstractNode<SimpleLink>
-{
-	
-}
